Fixed some minor issued that were breaking the simulation

+++ Crowd_pressure/src/main/java/simulation/model/Point.java
@@ -2,19 +2,19 @@ package simulation.model;
 
 public class Point {
 
-    private final int x;
-    private final int y;
+    private final double x;
+    private final double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
